pyproject.toml
[project]
name = "infotennis-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.30",
    "cryptography>=46.0.3",
    "duckdb>=1.4.3",
    "httpx>=0.28.1",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "prefect>=3.6.11",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "s3fs>=0.4.2",
    "tenacity>=9.1.2",
]
